Move balance task to correct namespace
