+++ ExampleApp/Tasks/Customers/Balance.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ExampleApp.Tasks
+namespace ExampleApp.Tasks.Customers
 {
     [Task("gcb", "Get Customer Balance")]
-    class Customer_Balance : BaseTask
+    class Balance : BaseTask
     {
         public override async Task Run()
         {
@@ -16,7 +16,7 @@ namespace ExampleApp.Tasks
 
             var rootRes = await Broker.GetRootAsync();
             var sourcesRes = await Broker.GetCustomerFundingSourcesAsync(new Uri($"{rootRes.Links["customers"].Href}/{input}"));
-            var balanceRes = await Broker.GetFundingSourceBalance(sourcesRes.Embedded.FundingSources.First(x => x.Type == "balance").Links["balance"].Href);
+            var balanceRes = await Broker.GetFundingSourceBalanceAsync(sourcesRes.Embedded.FundingSources.First(x => x.Type == "balance").Links["balance"].Href);
 
             WriteLine($" Balance: {balanceRes.Balance.Value} {balanceRes.Balance.Currency}");
         }
